requirements_tf.txt
# ===============================
# Core numerical stack (PIN FIRST)
# ===============================
numpy==1.24.3
scipy==1.10.1
pandas==2.1.4

# ===============================
# Classical ML
# ===============================
scikit-learn==1.3.0
imbalanced-learn==0.11.0
joblib==1.3.2
xgboost==1.7.6 

# ===============================
# Deep Learning (macOS)
# ===============================
tensorflow-macos==2.13.0
tensorflow-metal==1.0.0

# ===============================
# CV / NLP
# ===============================
opencv-python==4.8.0.74
Pillow<10
albumentations==1.3.1
nltk==3.9.0

# ===============================
# Visualization / Utils
# ===============================
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1
streamlit==1.40.2
python-dotenv==1.0.0


# ===============================
# API / Model Serving
# ===============================
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# ===============================
# Background / Infra
# ===============================
celery==5.3.1
redis==4.6.0
apscheduler==3.10.4

# ===============================
# Data / IO
# ===============================
openpyxl==3.1.2
requests==2.31.0
